x509: rename anchor module
